Refactor git commit verification into its own interface.
This pulls out git commit verification logic (e.g. everything but Rekor
lookups) into it's own interface + struct so that it can be more easily
configured with user settings.

The only new change in behavior is the inclusion of the system cert pool
in the default cert pool used for validation.

Signed-off-by: Billy Lynch <billy@chainguard.dev>